Moved `RatedVoteCase` into utilities_for_testing
